src/requirements.txt
torch
autogluon
Pillow